test(orchestrator): api/daemon-core/logger suites + timeout, panic, and WS-race fixes
Phase 4 of the testing plan. internal/api 0%->96.8% (httptest against the
shared tests/fixtures/internal-api wire fixtures, loaded at runtime with
sentinel substitution so Go and TS fail together on drift), pkg/retry
95.1%, internal/logger 0%->93.2%, container executor 6.5%->31% (non-Docker
paths; Docker cases stay behind the integration tag).

The daemon loop was extracted behavior-preserving from orchestrator.go
into an injectable daemonCore (core.go/spool.go/capture.go) behind
locally-defined interfaces the real api client/executor manager/logger/
metrics types satisfy as-is; extracted logic tested to 93.1%, race-clean.

Real bugs fixed, each with regression tests:
- Job timeouts never fired: the claim wire carries execution.timeout in
  nanoseconds but convertQueuedJob multiplied by time.Second again
  (int64 overflow: a 60s job got a ~147-year budget).
- A panic inside an executor killed the whole daemon, aborting every
  in-flight job; now contained (job reported FAILED on a detached
  context with its capability token, executor_panic metric, slot freed).
- WS log shipper: reconnect race lost messages (stale pump could close
  the new connection), shutdown raced queued sends into gorilla's
  concurrent-write panic, reconnectDelay data race, and Streamer.Stop()
  dropped the final flush-interval of buffered logs.

+++ apps/orchestrator/internal/api/types.go
@@ -40,7 +40,7 @@ type ExecutionConfig struct {
 	Script      *Script                `json:"script,omitempty"`
 	HTTP        *HTTPConfig            `json:"http,omitempty"`
 	Environment map[string]string      `json:"environment"`
-	Timeout     int                    `json:"timeout"` // seconds
+	Timeout     int64                  `json:"timeout"` // nanoseconds (serialized Go time.Duration)
 	Resources   *Resources             `json:"resources,omitempty"`
 	RetryPolicy *RetryPolicy           `json:"retryPolicy,omitempty"`
 	InputData   map[string]interface{} `json:"inputData,omitempty"`
